split GETAWAY & AUTOCORR3D
