+++ Code/GraphMol/Descriptors/CMakeLists.txt
@@ -1,6 +1,6 @@
 if(RDK_BUILD_DESCRIPTORS3D)
-  set(DESC3D_HDRS MolDescriptors3D.h PBF.h PMI.h RDF.h MORSE.h WHIM.h GETAWAY.h Data3Ddescriptors.h MolData3Ddescriptors.h) # auto3Dcorr.cpp)
-  set(DESC3D_SOURCES PBF.cpp PMI.cpp RDF.cpp MORSE.cpp WHIM.cpp GETAWAY.cpp Data3Ddescriptors.cpp MolData3Ddescriptors.cpp) # auto3Dcorr.h)
+  set(DESC3D_HDRS MolDescriptors3D.h PBF.h PMI.h RDF.h MORSE.h WHIM.h GETAWAY.h Data3Ddescriptors.h MolData3Ddescriptors.h AUTOCORR3D.h)
+  set(DESC3D_SOURCES PBF.cpp PMI.cpp RDF.cpp MORSE.cpp WHIM.cpp GETAWAY.cpp Data3Ddescriptors.cpp MolData3Ddescriptors.cpp AUTOCORR3D.cpp)
 endif(RDK_BUILD_DESCRIPTORS3D)
 
 rdkit_library(Descriptors
@@ -36,11 +36,10 @@ LINK_LIBRARIES  Descriptors FileParsers SmilesParse MolTransforms GraphMol DataS
 LINK_LIBRARIES  Descriptors FileParsers SmilesParse MolTransforms GraphMol DataStructs EigenSolvers RDGeneral RDGeometryLib ${RDKit_THREAD_LIBS} )
   rdkit_test(testGETAWAY testGETAWAY.cpp
 LINK_LIBRARIES  Descriptors FileParsers SmilesParse MolTransforms GraphMol DataStructs EigenSolvers RDGeneral RDGeometryLib ${RDKit_THREAD_LIBS} )
-#  rdkit_test(test3Dauto test3Dauto.cpp
-#LINK_LIBRARIES  Descriptors FileParsers SmilesParse MolTransforms GraphMol DataStructs EigenSolvers RDGeneral RDGeometryLib ${RDKit_THREAD_LIBS} )
+  rdkit_test(testAUTOCORR3D testAUTOCORR3D.cpp
+LINK_LIBRARIES  Descriptors FileParsers SmilesParse MolTransforms GraphMol DataStructs EigenSolvers RDGeneral RDGeometryLib ${RDKit_THREAD_LIBS} )
 
 endif(RDK_BUILD_DESCRIPTORS3D)
-
 
 
 add_subdirectory(Wrap)
