+++ Code/GraphMol/Descriptors/Data3Ddescriptors.cpp
@@ -13,10 +13,7 @@ double Data3Ddescriptors::ionpol[83]={1.208,0,0.479,0.828,0.737,1,1.291,1.209,1.
 double Data3Ddescriptors::rcov[83]={0.37,0,1.34,0.90,0.82,0.77,0.73,0.71,0,1.54,1.30,1.18,1.11,1.06,1.02,0.99,0,1.96,1.74,0,0,0,1.27,1.39,1.25,1.26,1.21,1.38,1.31,1.26,1.22,1.19,1.16,1.14,0,2.11,1.92,0,0,0,1.45,0,0,0,0,1.53,1.48,1.44,1.41,1.38,1.35,1.33,0,0,0,0,0,0,0,0,0,0,1.79,0,0,0,0,0,0,0,0,0,0,0,0,0,1.28,1.44,1.49,1.48,1.47, 1.46};
 
 
-Data3Ddescriptors::Data3Ddescriptors( )
-{
-
-}
+Data3Ddescriptors::Data3Ddescriptors(){}
 
 double* Data3Ddescriptors::getMW( ){
   	return mw;
